% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_options.R
\name{as_gdal_config_opts}
\alias{as_gdal_config_opts}
\alias{as_gdal_config_opts.default}
\alias{as_gdal_config_opts.gdal_config_opts}
\alias{as_gdal_config_opts.list}
\alias{as_gdal_config_opts.character}
\alias{as_gdal_config_opts.tbl_df}
\title{Coerce to GDAL Configuration Options}
\usage{
as_gdal_config_opts(x, ..., driver = NULL, call = rlang::caller_env())

\method{as_gdal_config_opts}{default}(x, ..., driver = NULL, call = rlang::caller_env())

\method{as_gdal_config_opts}{gdal_config_opts}(x, ..., driver = NULL, call = rlang::caller_env())

\method{as_gdal_config_opts}{list}(x, ..., driver = NULL, call = rlang::caller_env())

\method{as_gdal_config_opts}{character}(x, ..., driver = NULL, call = rlang::caller_env())

\method{as_gdal_config_opts}{tbl_df}(x, ..., driver = NULL, call = rlang::caller_env())
}
\arguments{
\item{x}{Object to coerce.}

\item{...}{Unused; for method extensibility.}

\item{driver}{Optional GDAL driver short name to attach.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A \code{\link[=gdal_config_opts]{gdal_config_opts()}} object.
}
\description{
Coerce a named list, a \code{KEY=VALUE} character vector, a driver-metadata tibble, or an existing
\code{gdal_config_opts} to the \code{\link[=gdal_config_opts]{gdal_config_opts()}} class.
}
