% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_gdalg.R
\name{as_gdalg}
\alias{as_gdalg}
\title{Coerce to GDALG}
\usage{
as_gdalg(x, ..., call = rlang::caller_env())
}
\arguments{
\item{x}{An object to coerce to class \code{"gdalg"}.}

\item{...}{Additional arguments passed to specific methods.}

\item{call}{The calling environment, used for error messages.}
}
\value{
An object of class \code{"gdalg"}.
}
\description{
Coerce to GDALG
}
