% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{avoid_backtraces}
\alias{avoid_backtraces}
\title{Avoid backtraces in examples}
\description{
This example should run first and set an option for the process that builds the example. By default, \code{pkgdown} builds
examples in a separate process.

This also produces a help page that is not linked from anywhere.
}
\examples{
options(rlang_backtrace_on_error = "none")
}
\keyword{internal}
