% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_available_ram}
\alias{check_available_ram}
\title{Check Available RAM}
\usage{
check_available_ram(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{The object to check.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
If the check passes, invisibly returns the provided \code{x} value. If the check fails, a condition error is thrown
indicating that the provided value exceeds available system RAM.
}
\description{
Checks that the provided value \code{x} (in bytes) does not exceed the available system RAM as returned by
\code{\link[=sys_available_ram]{sys_available_ram()}}. If the check fails, an error is thrown.
}
