% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_inherits}
\alias{check_inherits}
\alias{check_inherits2}
\alias{check_inherits_any}
\alias{check_inherits_all}
\title{Class Inheritence Checks}
\usage{
check_inherits(
  x,
  class,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

check_inherits2(
  x,
  class,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

check_inherits_any(
  x,
  classes,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

check_inherits_all(
  x,
  classes,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{The object to check.}

\item{class, classes}{The name of the class or classes to use during checking.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
If checks pass, invisibly returns the provided \code{x} object. If checks fail, a condition error is thrown.
}
\description{
These functions perform checks that assert the underlying class of objects passed to them.
\itemize{
\item \code{check_inherits()}: checks that object \code{x} is of class \code{class} using \code{\link[base:inherits]{base::inherits()}}
\item \code{check_inherits2()}: checks that object \code{x} is of class \code{class} using \code{\link[base:.class2]{base::.class2()}}
\item \code{check_inherits_any()}: checks that object \code{x} is at least one of the provided \code{classes} via \code{\link[rlang:inherits_any]{rlang::inherits_any()}}
\item \code{check_inherits_all()}: checks that object \code{x} is all of the provided \code{classes} via \code{\link[rlang:inherits_all]{rlang::inherits_all()}}
}

If validation fails for any of these functions, an error is thrown via \code{check_abort()} displaying a friendly
error message.
}
