% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_string}
\alias{check_string}
\title{Check String (Scalar)}
\usage{
check_string(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{The object to check.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
If checks pass, invisibly returns the provided \code{x} object. If checks fail, a condition error is thrown.
}
\description{
Checks the provided \code{x} is a scalar string via \code{\link[rlang:check_string]{rlang::check_string()}}.
}
