% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{checks}
\alias{checks}
\title{Check Functions}
\description{
Collection of various checking functions primarily used for incorporating argument validation checks for package
functions.

These check functions act as assertions, and will either return the provided objects invisibly, or throw exceptions.
}
\concept{checks}
\concept{utils}
