% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{.shared_returns_check}
\alias{.shared_returns_check}
\title{Check Returns}
\value{
If checks pass, invisibly returns to initially provided object \code{x}, otherwise a condition of class \code{check_error}
is thrown.
}
\description{
Returns for check_*() functions.
}
\keyword{internal}
