% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgb_validate.R
\name{fgb_validate_spatial_index_ram}
\alias{fgb_validate_spatial_index_ram}
\title{Validate Spatial Index RAM Requirement}
\usage{
fgb_validate_spatial_index_ram(fgb_dsn, force = FALSE, quiet = FALSE)
}
\arguments{
\item{fgb_dsn}{The data source name (DSN) of the FlatGeobuf file to validate.}

\item{force}{Logical indicating whether to force the feature count (default: \code{FALSE}).}

\item{quiet}{Logical indicating whether to suppress success messages (default: \code{FALSE}).}
}
\value{
Returns \code{TRUE} if sufficient RAM is available to build the spatial index, and \code{FALSE} otherwise.
Also provides informative messages about the RAM requirements and availability.
}
\description{
Validates if the available RAM is sufficient to build a spatial index for a FlatGeobuf file based on the number of
features and an estimated RAM requirement of 83 bytes per feature.
}
