% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_config.R
\name{gdal_config}
\alias{gdal_config}
\title{GDAL Configuration}
\usage{
gdal_config(...)
}
\value{
gdal_config
}
\description{
Creates a session-wide, global GDAL Configuration object of class \code{"gdal_config"}.
}
