% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{gdal_config_links}
\alias{gdal_config_links}
\title{GDAL Configuration Links}
\description{
Links to the official GDAL Configuration Documentation.
\itemize{
\item \href{https://gdal.org/en/stable/user/configoptions.html}{GDAL Configuration Options}
\itemize{
\item \href{https://gdal.org/en/stable/user/configoptions.html#setting-configuration-options}{Setting Configuration Options}
}
}
}
\seealso{
\code{\link[=gdal_config]{gdal_config()}}
}
