% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalvector-package.R
\docType{package}
\name{gdalvector-package}
\alias{gdalvector}
\alias{gdalvector-package}
\title{gdalvector: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
\keyword{internal}
