% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_predicates.R
\name{is_vsi_path}
\alias{is_vsi_path}
\title{Is VSI Path}
\usage{
is_vsi_path(x)
}
\arguments{
\item{x}{Character string to check.}
}
\value{
Logical indicating if the path starts with a valid VSI prefix (i.e. \verb{/vsicurl/} or \verb{/vsizip/}).
}
\description{
Check if a Path is a GDAL Virtual File System (VSI) Path or URL.
}
\examples{
is_vsi_path("/vsizip/data.zip")      # TRUE
is_vsi_path("/vsicurl/data.geojson") # TRUE
is_vsi_path("data.geojson")          # FALSE

}
