% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_remote.R
\name{local_hash}
\alias{local_hash}
\title{Get the Hash of a Local File}
\usage{
local_hash(path, algo = "md5")
}
\arguments{
\item{path}{Character string specifying the path to the local file.}

\item{algo}{Character string specifying hash algorithm ("md5", "sha1", "sha256", "sha512").}
}
\value{
A character string representing the hash of the local file.
}
\description{
Retrieves the hash of a local file from its path.
}
\examples{
\dontrun{
path <- "data-raw/cache/tiger/GENZ2024/shp/cb_2024_us_state_20m.zip"
local_hash(path)
}
}
