% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_remote.R
\name{ping}
\alias{ping}
\title{Ping}
\usage{
ping(url, timeout = 5L)
}
\arguments{
\item{url}{The remote URL}

\item{timeout}{Timeout in seconds. Defaults to \code{5L}.}
}
\value{
\code{TRUE} if successfully pinged, \code{FALSE} otherwise.
}
\description{
Pings a remote URL.
}
