% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_predicates.R
\name{predicates}
\alias{predicates}
\alias{is_int64}
\title{Predicates}
\usage{
is_int64(x)
}
\description{
A collection of predicate functions for checking various conditions, types, and properties of objects.
These functions return logical values (TRUE or FALSE) based on the input provided.
}
\section{Functions}{

\subsection{Types}{
\itemize{
\item \code{is_int64(x)}: Check if \code{x} is of type \code{integer64}.
}
}

\subsection{Virtual File System (VSI)}{
\itemize{
\item \code{is_vsi_path(x)}: Check if a path is a GDAL Virtual File System (VSI) path or URL.
}
}
}

