% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_remote.R
\name{remote_exists}
\alias{remote_exists}
\title{Check if a Remote File Exists at a Given URL}
\usage{
remote_exists(url)
}
\arguments{
\item{url}{Character string specifying the URL of the remote file.}
}
\value{
A logical value: \code{TRUE} if the remote file exists (HTTP status 200), \code{FALSE} otherwise.
}
\description{
Determines whether a remote file exists by sending a \code{HEAD} request to the specified
URL and checking the HTTP status code.
}
\examples{
\dontrun{
url <- "https://www2.census.gov/geo/tiger/GENZ2024/shp/cb_2024_us_state_20m.zip"
remote_exists(url)
}
}
