% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_remote.R
\name{remote_hash}
\alias{remote_hash}
\title{Get the Hash of a Remote File from its URL}
\usage{
remote_hash(url, algo = "md5")
}
\arguments{
\item{url}{Character string specifying the URL of the remote file.}

\item{algo}{Character string specifying hash algorithm ("md5", "sha1", "sha256", "sha512").}
}
\value{
A character string representing the hash of the remote file.
}
\description{
Retrieves the hash of a remote file from its URL.
}
\examples{
\dontrun{
url <- "https://www2.census.gov/geo/tiger/GENZ2024/shp/cb_2024_us_state_20m.zip"
remote_hash(url)
}
}
