% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_remote.R
\name{remote_head}
\alias{remote_head}
\title{Perform a \code{HEAD} HTTP Request for a Remote URL}
\usage{
remote_head(url)
}
\arguments{
\item{url}{Character string specifying the URL to send the \code{HEAD} request to.}
}
\value{
A list containing:
\itemize{
\item \code{request}: The \code{\link[httr2:request]{httr2::request()}}
\item \code{response}: The \code{\link[httr2:response]{httr2::response()}}
\item \code{headers}: The \code{\link[httr2:resp_headers]{httr2::resp_headers()}} from the response
}
}
\description{
Sends a \code{HEAD} request to the specified URL and retrieves the response headers.

This function is useful for checking the existence of a resource or retrieving metadata
without downloading the entire content.
}
\examples{
\dontrun{
url <- "https://www2.census.gov/geo/tiger/GENZ2024/shp/cb_2024_us_state_20m.zip"
head_info <- remote_head(url)
print(head_info$headers)
}
}
