% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_remote.R
\name{remote_last_modified}
\alias{remote_last_modified}
\alias{local_last_modified}
\title{Get Local & Remote Resources Last-Modified Timestamp}
\usage{
remote_last_modified(url)

local_last_modified(path)
}
\arguments{
\item{url}{Character string specifying the URL of the remote file.}

\item{path}{Path to local file to get the last modified timestamp for.}
}
\value{
\itemize{
\item \code{remote_last_modified()}: \code{POSIXct} datetime representing the \code{Last-Modified} header timestamp, or \code{NA} if not provided by the
server.
\item \code{local_last_modified()}: \code{POSIXct} datetime of the file's last modification.
}
}
\description{
These functions derive timestamps for local and remote resources.
\itemize{
\item \code{remote_last_modified()}: Parses the \code{Last-Modified} HTTP response header as the timestamp.
Returns \code{NA} if the header is not available.
\item \code{local_last_modified()}: Local file last modified timestamp.
}
}
\examples{
\dontrun{
url <- "https://www2.census.gov/geo/tiger/GENZ2024/shp/cb_2024_us_state_20m.zip"
remote_last_modified(url)

path <- "data-raw/cache/tiger/GENZ2024/shp/cb_2024_us_state_20m.zip"
local_last_modified(path)
}
}
