% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_remote.R
\name{remote_size}
\alias{remote_size}
\title{Get the Size of a Remote File from its URL}
\usage{
remote_size(url)
}
\arguments{
\item{url}{Character string specifying the URL of the remote file.}
}
\value{
A numeric value representing the size of the remote file in bytes.
}
\description{
Retrieves the size of a remote file by sending a \code{HEAD} request to the specified
URL and extracting the \code{Content-Length} header.
}
\examples{
\dontrun{
url <- "https://www2.census.gov/geo/tiger/GENZ2024/shp/cb_2024_us_state_20m.zip"
remote_size(url)
}
}
