% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sql.R
\name{sql_in}
\alias{sql_in}
\title{Build a SQL \code{IN} Clause}
\usage{
sql_in(field, values, negate = FALSE)
}
\arguments{
\item{field}{Character. The column name.}

\item{values}{Values to include in the clause.}

\item{negate}{Logical. If \code{TRUE}, use \verb{NOT IN}. Default is \code{FALSE}.}
}
\value{
A character string containing the SQL clause.
}
\description{
Constructs a SQL \code{IN} or \verb{NOT IN} clause for a given field and set of values.
}
\examples{
non_conus_state_fips <- c("02", "15", "60", "66", "69", "72", "74", "78")
sql_in(field = "STATEFP", values = non_conus_state_fips, negate = TRUE)
}
\keyword{internal}
