% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sql.R
\name{sql_where_valid_geom}
\alias{sql_where_valid_geom}
\title{Build a SQL \code{ST_IsValid} WHERE Clause}
\usage{
sql_where_valid_geom(geom_col, negate = FALSE)
}
\arguments{
\item{geom_col}{Character. The name of the geometry column.}

\item{negate}{Logical. If \code{TRUE}, filters for \emph{invalid} geometries. Default is \code{FALSE}.}
}
\value{
A character string containing the SQL expression.
}
\description{
Constructs a SQL \code{ST_IsValid()} expression for filtering geometries.
}
\examples{
sql_where_valid_geom("geom")
sql_where_valid_geom("geom", negate = TRUE)
}
