% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_num_cpus}
\alias{sys_num_cpus}
\title{\code{sys_num_cpus} - System Number of CPUs}
\usage{
sys_num_cpus()
}
\value{
Integer representing the number of CPU cores available to the R session.
}
\description{
Get the number of CPU cores available on the current machine using \code{\link[gdalraster:get_num_cpus]{gdalraster::get_num_cpus()}},
which calls the internal GDAL C++ library function \code{GDALGetCPUs()}.
}
\details{
This method is more robust than \code{parallel::detectCores()} because it accounts for CPU
affinity, container limits, and other environmental restrictions that may cap the processing
pools actually available to the R session.

However, on a standard unconstrained desktop machine, it will return the same value as
\code{parallel::detectCores(logical = TRUE)} because both report the total logical processing channels.
\itemize{
\item \strong{Number of CPUs (GDAL):} \code{gdalraster::get_num_cpus()} queries the C++ backend to see how many logical
execution slots (hardware threads) are exposed by the operating system.
\item \strong{Number of Cores (parallel):} \code{parallel::detectCores(logical = TRUE)} targets the virtual threads generated
by Hyper-Threading (Intel) or SMT (AMD). Conversely, running \code{parallel::detectCores(logical = FALSE)} attempts
to return only the count of independent physical cores on the processor.
}
}
\examples{
\dontrun{
sys_num_cpus()
}
}
