% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_os}
\alias{sys_os}
\title{\code{sys_os} - System OS Name}
\usage{
sys_os()
}
\value{
Character string resulting from \code{Sys.info()[["sysname"]]}, which will be one of c("windows", "linux", "darwin", etc.)
depending on the system.
}
\description{
Get the current machine's operating system name.
}
\examples{
\dontrun{
sys_os()
}
}
