% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_path}
\alias{sys_path}
\title{\code{sys_path} - System PATH}
\usage{
sys_path(filter = NULL)
}
\arguments{
\item{filter}{Optional character string. If provided, only paths containing this string will be returned.}
}
\value{
Character vector of paths from the system's \code{PATH} environment variable, split by the appropriate path separator
for the operating system. If \code{filter} is provided, only paths containing the filter string are included in the
returned vector. If no paths match the filter, an empty character vector is returned: \code{character(0)}.

If \code{filter} is provided, only paths containing the filter string are returned.
}
\description{
Get the Current Machine's PATH Environment Variable as a Character Vector
}
\examples{
\dontrun{
sys_path()
}
}
