% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_pid}
\alias{sys_pid}
\title{\code{sys_pid} - System Process ID}
\usage{
sys_pid()
}
\value{
Integer representing the current process ID.
}
\description{
Get the current process ID of the R session.
}
\examples{
\dontrun{
sys_pid()
}
}
