% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_platform}
\alias{sys_platform}
\title{\code{sys_platform} - System Platform}
\usage{
sys_platform()
}
\value{
Character string resulting from \code{.Platform$OS.type}
}
\description{
Get the current machine's platform (operating system "family")
}
\examples{
\dontrun{
sys_platform()
}
}
