% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_system.R
\name{sys_which}
\alias{sys_which}
\title{\code{sys_which} - System \code{which}}
\usage{
sys_which(x, winslash = "/", ...)
}
\arguments{
\item{x}{Passed to \code{Sys.which()} \code{names} argument.}

\item{winslash}{the separator to be used on Windows -- ignored
    elsewhere.  Must be one of \code{c("/", "\\\\")}.}

\item{...}{
  Arguments passed on to \code{\link[base:normalizePath]{base::normalizePath}}
  \describe{
    \item{\code{path}}{character vector of file paths.}
    \item{\code{mustWork}}{logical: if \code{TRUE} then an error is given if the result
    cannot be determined; if \code{NA} then a warning.}
  }}
}
\value{
Character vector of paths, if found. If not found returns \code{NULL} instead of \code{""}.
}
\description{
Lightweight, convenience wrapper around \code{\link[base:Sys.which]{base::Sys.which()}} and \code{\link[base:normalizePath]{base::normalizePath()}}.
}
\examples{
\dontrun{
sys_which("gdal")
}
}
