% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_azure}
\alias{vsi_azure}
\title{Build a \verb{/vsiaz/} Path}
\usage{
vsi_azure(container, blob)
}
\arguments{
\item{container}{Azure Blob container name.}

\item{blob}{Blob path within the container.}
}
\value{
Character scalar.
}
\description{
Build a \verb{/vsiaz/} Path
}
