% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_curl}
\alias{vsi_curl}
\title{Wrap a URL with \verb{/vsicurl/}}
\usage{
vsi_curl(url)
}
\arguments{
\item{url}{URL to wrap.}
}
\value{
Character vector of \verb{/vsicurl/} paths.
}
\description{
Wrap a URL with \verb{/vsicurl/}
}
