% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_exists}
\alias{vsi_exists}
\title{Test Whether a VSI Path Exists}
\usage{
vsi_exists(path)
}
\arguments{
\item{path}{Path to test.}
}
\value{
Logical vector.
}
\description{
Test Whether a VSI Path Exists
}
