% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_from_uri}
\alias{vsi_from_uri}
\title{Convert a URI to a GDAL VSI Path}
\usage{
vsi_from_uri(uri)
}
\arguments{
\item{uri}{URI or URL.}
}
\value{
Character vector of VSI paths.
}
\description{
Convert a URI to a GDAL VSI Path
}
