% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_handlers}
\alias{vsi_handlers}
\title{List VSI Handlers Used by a Path}
\usage{
vsi_handlers(path)
}
\arguments{
\item{path}{VSI path.}
}
\value{
Character vector of handler names.
}
\description{
List VSI Handlers Used by a Path
}
