% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_meta}
\alias{vsi_meta}
\title{Get VSI File Metadata}
\usage{
vsi_meta(path, domain = "HEADERS")
}
\arguments{
\item{path}{Path to inspect.}

\item{domain}{Metadata domain, typically \code{"HEADERS"} or \code{"ZIP"}.}
}
\value{
Named character vector of metadata entries.
}
\description{
Get VSI File Metadata
}
