% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_path}
\alias{vsi_path}
\title{Convert a Path to a GDAL VSI Path}
\usage{
vsi_path(path, ...)
}
\arguments{
\item{path}{A path, URL, or object with a \code{vsi_path()} method.}

\item{...}{Passed to methods.}
}
\value{
Character vector of VSI paths.
}
\description{
Convert a Path to a GDAL VSI Path
}
