% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_size}
\alias{vsi_size}
\title{Get the Size of a VSI Path}
\usage{
vsi_size(path)
}
\arguments{
\item{path}{Path to inspect.}
}
\value{
The size of the file/object in bytes, formatted using \code{\link[rlang:as_bytes]{rlang::as_bytes()}}.
}
\description{
Get the Size of a VSI Path
}
