% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_strip}
\alias{vsi_strip}
\title{Strip the Outermost VSI Handler}
\usage{
vsi_strip(path, recurse = TRUE)
}
\arguments{
\item{path}{Path to strip.}
}
\value{
Character vector with the outermost VSI handler removed.
}
\description{
Strip the Outermost VSI Handler
}
