% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_type}
\alias{vsi_type}
\title{Get the Type of a VSI Path}
\usage{
vsi_type(path)
}
\arguments{
\item{path}{Path to inspect.}
}
\value{
Character vector of path types.
}
\description{
Get the Type of a VSI Path
}
