% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_zip}
\alias{vsi_zip}
\title{Wrap a Path with \verb{/vsizip/}}
\usage{
vsi_zip(path)
}
\arguments{
\item{path}{Path to a ZIP archive.}
}
\value{
Character vector of \verb{/vsizip/} paths.
}
\description{
Wrap a Path with \verb{/vsizip/}
}
