% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_vsi.R
\name{vsi_zip_curl}
\alias{vsi_zip_curl}
\title{Build a \verb{/vsizip//vsicurl/} Path}
\usage{
vsi_zip_curl(url, inner = NULL)
}
\arguments{
\item{url}{Remote ZIP URL.}

\item{inner}{Optional inner file path within the ZIP archive.}
}
\value{
Character scalar.
}
\description{
Build a \verb{/vsizip//vsicurl/} Path
}
