% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{xml_parse_gdal_driver_config_opts}
\alias{xml_parse_gdal_driver_config_opts}
\title{Parse GDAL Driver Configuration Options from XML}
\usage{
xml_parse_gdal_driver_config_opts(
  xml,
  scope = "all",
  driver = NULL,
  type = "config",
  call = rlang::caller_env()
)
}
\arguments{
\item{xml}{The XML document to parse, typically obtained from a GDAL driver's documentation page.}

\item{scope}{The scope of the options being parsed, such as "vector", "raster", or "all".
This is used to categorize the options based on their applicability to different data types. Defaults to "all".}

\item{driver}{The name of the GDAL driver for which the options are being parsed.
This is used for labeling purposes in the resulting tibble. Defaults to \code{NULL}.}

\item{call}{The calling function, used for error handling and messaging.}

\item{opt_type}{The type of options being parsed, such as "config", "open", or "creation".
This is used for labeling purposes in the resulting tibble. Defaults to "config".}
}
\value{
A tibble with columns for \code{name}, \code{description}, \code{scope}, \code{default}, and \code{values}
}
\description{
Parses the configuration options for a GDAL driver from the provided XML document. This function
is specifically designed to extract the configuration options listed in the "Configuration Options" section of a GDAL
driver's documentation page, which is typically structured as an unordered list (\verb{<ul>}) with list
items (\verb{<li>}) containing the option details.

The function looks for the first \verb{<ul>} element within the section with \code{id="configuration-options"} and extracts
the option name, description, default value, and possible values (if specified in brackets).
The resulting data is returned as a tibble with columns for \code{name}, \code{description}, \code{scope}, \code{default}, and \code{values}
(a list-column containing character vectors of possible values).
}
