% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_xml.R
\name{xml_parse_gdal_options}
\alias{xml_parse_gdal_options}
\title{Parse XML for GDAL Options}
\usage{
xml_parse_gdal_options(
  xml,
  driver = NA_character_,
  type = c("config", "open", "creation"),
  sub_type = NA_character_,
  scope = c("vector", "raster", "all"),
  call = rlang::caller_env()
)
}
\arguments{
\item{xml}{The XML to be parsed. Must be either a valid XML character string or an \code{xml2::xml_document} object.}

\item{driver, opt_type}{(Optional) Additional values that can be added to the resulting \code{\link[tibble:tibble]{tibble::tibble()}}. Useful
for when merging options across multiple drivers or option types. Defaults to \code{NULL} and the columns
will only be included in the output when provided. These values are also not properly validated against
the GDAL drivers or option types, so they should be used with caution and primarily for internal use.}

\item{scope}{(Optional) The scope of the options being parsed in terms of the supported data types, i.e. \code{vector}
vs. \code{raster} or both. For example, \code{GPKG} has \code{scope}s defined for \code{"vector"}, \code{"raster"}, \code{"raster,vector"}, and
\code{NA}. Use this argument to filter out for only the options for a specific scope. Will always include the \code{NA} scoped
options, as those are not explicitly defined for a specific scope and are likely applicable to all scopes. Defaults to \code{NULL}
and will not filter for any specific scope, returning all options regardless of scope. Provided value
must be one of \code{"vector"}, \code{"raster"}, or \code{"all"} if not \code{NULL}. Note that the scope values are not standardized
across all GDAL drivers, so use with caution and always check the resulting tibbles for the expected scope values
when working with drivers supporting both raster and vector data.}

\item{call}{The calling function, used for error handling and messaging.}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} with fields for \code{name}, \code{description}, \code{scope}, \code{type}, \code{default}, and
\code{values} (a list-column with a character vector for the possible values).

Additional fields for \code{driver} and \code{opt_type} may be included if those parameters are provided.
}
\description{
Parses the XML from GDAL driver metadata into structured \code{tbl_df} tibbles.

The function is meant to be flexible enough to be able to parse any of the possible
XML metadata structures from GDAL's registered (vector) driver's metadata:
\code{DMD_OPENOPTIONLIST}, \code{DMD_CREATIONOPTIONLIST}, and \code{DS_LAYER_CREATIONOPTIONLIST}.
}
\examples{
# parse DS_LAYER_CREATIONOPTIONLIST
gdalraster::gdal_get_driver_md("GPKG", mdi_name = "DS_LAYER_CREATIONOPTIONLIST") |>
  xml_parse_gdal_options()

# parse DMD_CREATIONOPTIONLIST
gdalraster::gdal_get_driver_md("GPKG", mdi_name = "DMD_CREATIONOPTIONLIST") |>
  xml_parse_gdal_options()

# parse DMD_OPENOPTIONLIST
gdalraster::gdal_get_driver_md("GPKG", mdi_name = "DMD_OPENOPTIONLIST") |>
  xml_parse_gdal_options()
}
