% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalvector-package.R
\docType{package}
\name{gdalvector-package}
\alias{gdalvector}
\alias{gdalvector-package}
\title{\code{gdalvector} Package}
\description{
The \code{gdalvector} package provides tools for working with vector data using modern GDAL bindings in R. It builds
on top of the \code{gdalraster} package, but focuses on vector data sources and modern cloud-native formats and
specifications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/gdalvector}
  \item \url{http://docs.jimbrig.com/gdalvector/}
  \item Report bugs at \url{https://github.com/jimbrig/gdalvector/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
\keyword{internal}
