% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_predicates.R
\name{is_vsi_path}
\alias{is_vsi_path}
\title{Check if Virtual File System (VSI)}
\usage{
is_vsi_path(x)
}
\arguments{
\item{x}{Character string to check. Should represent a path or URL (or in rare cases, a connection string).}
}
\value{
Logical indicating if the path starts with a valid VSI prefix.
}
\description{
Check if a provided path or URL uses \href{https://gdal.org/en/stable/user/virtual_file_systems.html}{GDAL's Virtual File System (VSI)}.

Virtual File System's include prefixes such as \verb{/vsistdin/}, \verb{/vsistdout/}, \verb{/vsimem/}, etc.
}
\examples{
is_vsi_path("/vsizip/data.zip")      # TRUE
is_vsi_path("/vsicurl/data.geojson") # TRUE
is_vsi_path("data.geojson")          # FALSE

}
