% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_opts.R
\name{shp_creation_opts}
\alias{shp_creation_opts}
\title{ESRI Shapefile Creation Options}
\usage{
shp_creation_opts(
  spatial_index = NULL,
  encoding = NULL,
  resize = NULL,
  shpt = NULL,
  two_gb_limit = NULL,
  auto_repack = NULL,
  dbf_date_last_update = NULL,
  dbf_eof_char = NULL,
  .set_defaults = FALSE
)
}
\arguments{
\item{spatial_index}{Value for \code{SPATIAL_INDEX} (logical -> \code{"YES"}/\code{"NO"}).}

\item{encoding}{Value for \code{ENCODING} (DBF encoding).}

\item{resize}{Value for \code{RESIZE} (logical -> \code{"YES"}/\code{"NO"}); resize fields to optimal size.}

\item{shpt}{Value for \code{SHPT} (shape type, e.g. \code{"POLYGON"}).}

\item{two_gb_limit}{Value for \verb{2GB_LIMIT} (logical -> \code{"YES"}/\code{"NO"}); restrict \code{.shp}/\code{.dbf} to 2 GB.}

\item{auto_repack}{Value for \code{AUTO_REPACK} (logical -> \code{"YES"}/\code{"NO"}).}

\item{dbf_date_last_update}{Value for \code{DBF_DATE_LAST_UPDATE} (\code{YYYY-MM-DD}).}

\item{dbf_eof_char}{Value for \code{DBF_EOF_CHAR} (logical -> \code{"YES"}/\code{"NO"}).}

\item{.set_defaults}{Logical. If \code{TRUE}, options left unset (\code{NULL}) are filled with the driver's
GDAL metadata defaults (via the relevant \verb{gdal_driver_get_*_opts_defaults()}); user-supplied
values always take precedence. Defaults to \code{FALSE}.}
}
\value{
A layer-level \code{\link[=gdal_creation_opts]{gdal_creation_opts()}} object for the \verb{ESRI Shapefile} driver.
}
\description{
Construct a layer-level \code{\link[=gdal_creation_opts]{gdal_creation_opts()}} object for the \verb{ESRI Shapefile} driver.
}
\examples{
shp_creation_opts(spatial_index = TRUE, encoding = "UTF-8")
}
\seealso{
\code{\link[=shp_open_opts]{shp_open_opts()}}, \code{\link[=gdal_creation_opts]{gdal_creation_opts()}}
\itemize{
\item \href{http://shapelib.maptools.org/}{Shapefile Home Page}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html}{Shapefile GDAL Driver}
\itemize{
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#driver-capabilities}{Shapefile GDAL Capabilities}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#open-options}{Shapefile GDAL Open Options}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#layer-creation-options}{Shapefile GDAL Layer Creation Options}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#configuration-options}{Shapefile GDAL Configuration Options}
}
\item \href{http://shapelib.maptools.org/}{Shapefile C Library}
\item \href{http://dl.maptools.org/dl/shapelib/shapefile.pdf}{ESRI Shapefile Technical Description (PDF)}
\item \href{http://shapelib.maptools.org/shp_api.html}{Shapefile .SHP File API}
\item \href{http://shapelib.maptools.org/dbf_api.html}{Attribute .DBF File API}
\item \href{https://www.clicketyclick.dk/databases/xbase/format/}{Xbase File Format Description}
\item \href{http://shapelib.maptools.org/codepage.html}{Shapelib Code Page}
}
}
