% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_opts.R
\name{shp_open_opts}
\alias{shp_open_opts}
\title{ESRI Shapefile Open Options}
\usage{
shp_open_opts(
  encoding = NULL,
  dbf_date_last_update = NULL,
  adjust_type = NULL,
  adjust_geom_type = NULL,
  auto_repack = NULL,
  dbf_eof_char = NULL,
  .set_defaults = FALSE
)
}
\arguments{
\item{encoding}{Value for \code{ENCODING} (override DBF encoding).}

\item{dbf_date_last_update}{Value for \code{DBF_DATE_LAST_UPDATE} (\code{YYYY-MM-DD}).}

\item{adjust_type}{Value for \code{ADJUST_TYPE} (logical -> \code{"YES"}/\code{"NO"}).}

\item{adjust_geom_type}{Value for \code{ADJUST_GEOM_TYPE} (e.g. \code{"FIRST_SHAPE"}).}

\item{auto_repack}{Value for \code{AUTO_REPACK} (logical -> \code{"YES"}/\code{"NO"}).}

\item{dbf_eof_char}{Value for \code{DBF_EOF_CHAR} (logical -> \code{"YES"}/\code{"NO"}).}

\item{.set_defaults}{Logical. If \code{TRUE}, options left unset (\code{NULL}) are filled with the driver's
documented GDAL metadata defaults (via the relevant \verb{gdal_vector_driver_*_opts_defaults()});
user-supplied values always take precedence. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[=gdal_open_opts]{gdal_open_opts()}} object for the \verb{ESRI Shapefile} driver.
}
\description{
Construct a \code{\link[=gdal_open_opts]{gdal_open_opts()}} object for the \verb{ESRI Shapefile} driver.
}
\examples{
shp_open_opts(encoding = "UTF-8", auto_repack = TRUE)
}
\seealso{
\code{\link[=shp_config_opts]{shp_config_opts()}}, \code{\link[=shp_creation_opts]{shp_creation_opts()}}, \code{\link[=gdal_open_opts]{gdal_open_opts()}}
\itemize{
\item \href{http://shapelib.maptools.org/}{Shapefile Home Page}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html}{Shapefile GDAL Driver}
\itemize{
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#driver-capabilities}{Shapefile GDAL Capabilities}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#open-options}{Shapefile GDAL Open Options}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#layer-creation-options}{Shapefile GDAL Layer Creation Options}
\item \href{https://gdal.org/en/stable/drivers/vector/shapefile.html#configuration-options}{Shapefile GDAL Configuration Options}
}
\item \href{http://shapelib.maptools.org/}{Shapefile C Library}
\item \href{http://dl.maptools.org/dl/shapelib/shapefile.pdf}{ESRI Shapefile Technical Description (PDF)}
\item \href{http://shapelib.maptools.org/shp_api.html}{Shapefile .SHP File API}
\item \href{http://shapelib.maptools.org/dbf_api.html}{Attribute .DBF File API}
\item \href{https://www.clicketyclick.dk/databases/xbase/format/}{Xbase File Format Description}
\item \href{http://shapelib.maptools.org/codepage.html}{Shapelib Code Page}
}
}
